chore(wfs): highlight shared CollectionPropertyType in pivot model (refs #37) and review WFS conversions (refs #46)
